refactor: adjust Mellon.Multitenant.Azure project - to work with the new editorconfig and update imports - update MultiTenantExtensions to use the namespace Microsoft.Extensions.DependencyInjection (#breaking change) - update the packages  - Azure.Identity from `1.10.4` to `1.12.1`  - Microsoft.Extensions.Configuration.AzureAppConfiguration from `7.0.0` to `7.3.0`
