+++ src/Mellon.MultiTenant.Azure/AzureMultiTenantOptions.cs
@@ -1,10 +1,10 @@
-﻿using Microsoft.Extensions.Configuration.AzureAppConfiguration;
+﻿namespace Mellon.MultiTenant.Azure;
 
-namespace Mellon.MultiTenant.Azure;
+using Microsoft.Extensions.Configuration.AzureAppConfiguration;
 
 public class AzureMultiTenantOptions
 {
-    public string AzureAppConfigurationConnectionString { get; set; }
+	public string AzureAppConfigurationConnectionString { get; set; }
 
-    public Func<IServiceProvider, string, Action<AzureAppConfigurationOptions>> AzureAppConfigurationOptions { get; set; }
+	public Func<IServiceProvider, string, Action<AzureAppConfigurationOptions>> AzureAppConfigurationOptions { get; set; }
 }
