+++ src/Mellon.MultiTenant.Azure/Mellon.MultiTenant.Azure.csproj
@@ -1,39 +1,39 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Title>Mellon.MultiTenant.Azure</Title>
-    <PackageId>Mellon-MultiTenant-Azure</PackageId>
-    <Authors>1berto and Naga</Authors>
-    <Company>PubDev</Company>
-    <Description>A Package to create Multi-Tenant Applications using NET</Description>
-    <PackageLicenseExpression>MIT</PackageLicenseExpression>
-    <PackageProjectUrl>https://pub-dev.github.io/Mellon.MultiTenant/</PackageProjectUrl>
-    <RepositoryUrl>https://github.com/1bberto/Mellon.MultiTenant</RepositoryUrl>
-    <RepositoryType>git</RepositoryType>
-    <PackageRequireLicenseAcceptance>false</PackageRequireLicenseAcceptance>
-    <PackageIcon>logo.png</PackageIcon>
-    <PackageReadmeFile>readme.md</PackageReadmeFile>
-    <PublishRepositoryUrl>true</PublishRepositoryUrl>
-    <EmbedUntrackedSources>false</EmbedUntrackedSources>
-    <IncludeSymbols>true</IncludeSymbols>
-    <SymbolPackageFormat>snupkg</SymbolPackageFormat>
-  </PropertyGroup>
+	<PropertyGroup>
+		<TargetFramework>net8.0</TargetFramework>
+		<ImplicitUsings>enable</ImplicitUsings>
+		<Title>Mellon.MultiTenant.Azure</Title>
+		<PackageId>Mellon-MultiTenant-Azure</PackageId>
+		<Authors>1berto and Naga</Authors>
+		<Company>PubDev</Company>
+		<Description>A Package to create Multi-Tenant Applications using NET</Description>
+		<PackageLicenseExpression>MIT</PackageLicenseExpression>
+		<PackageProjectUrl>https://pub-dev.github.io/Mellon.MultiTenant/</PackageProjectUrl>
+		<RepositoryUrl>https://github.com/1bberto/Mellon.MultiTenant</RepositoryUrl>
+		<RepositoryType>git</RepositoryType>
+		<PackageRequireLicenseAcceptance>false</PackageRequireLicenseAcceptance>
+		<PackageIcon>logo.png</PackageIcon>
+		<PackageReadmeFile>readme.md</PackageReadmeFile>
+		<PublishRepositoryUrl>true</PublishRepositoryUrl>
+		<EmbedUntrackedSources>false</EmbedUntrackedSources>
+		<IncludeSymbols>true</IncludeSymbols>
+		<SymbolPackageFormat>snupkg</SymbolPackageFormat>
+	</PropertyGroup>
 
-  <ItemGroup>
-    <None Include="..\..\readme.md" Pack="true" PackagePath="\" />
-    <None Include="..\..\logo.png" Pack="true" PackagePath="\" />
-  </ItemGroup>
+	<ItemGroup>
+		<None Include="..\..\readme.md" Pack="true" PackagePath="\" />
+		<None Include="..\..\logo.png" Pack="true" PackagePath="\" />
+	</ItemGroup>
 
-  <ItemGroup>
-    <PackageReference Include="Azure.Identity" Version="1.10.4" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.AzureAppConfiguration" Version="7.0.0" />
-    <PackageReference Include="Microsoft.SourceLink.GitHub" Version="8.0.0" PrivateAssets="All" />
-  </ItemGroup>
+	<ItemGroup>
+		<PackageReference Include="Azure.Identity" />
+		<PackageReference Include="Microsoft.Extensions.Configuration.AzureAppConfiguration" />
+		<PackageReference Include="Microsoft.SourceLink.GitHub" />
+	</ItemGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\Mellon.MultiTenant.Base\Mellon.MultiTenant.Base.csproj" />
-  </ItemGroup>
+	<ItemGroup>
+		<ProjectReference Include="..\Mellon.MultiTenant.Base\Mellon.MultiTenant.Base.csproj" />
+	</ItemGroup>
 
 </Project>
